Consolidate ECS files - remove confusing _full versions
- Merged game_world_full.cpp into game_world.cpp
- Merged inventory_system_full.cpp into inventory_system.cpp
- Updated CMakeLists.txt to use regular files
- Removed _full files to avoid confusion
- Debug output still shows player_id=0 issue to fix
